+++ CHANGELOG.md
@@ -1,16 +1,21 @@
-# Version 1.0.0
+### Unreleased
+* enhancements
+  * `discard_all` and `undiscard_all` now return affected records
+  * Add `discard!` and `undiscard!`
+  
+### Version 1.0.0
 Release date: 2018-03-16
 
 * Add undiscard callbacks and `.undiscard_all`
 
-# Version 0.2.0
+### Version 0.2.0
 Release date: 2017-11-22
 
 * Add `.discard_all`
 * Add `undiscarded` scope
 * Add callbacks
 
-# Version 0.1.0
+### Version 0.1.0
 Release date: 2017-04-28
 
 * Initial version!